Fix Travis errors.

+++ scram_test.go
@@ -5,9 +5,13 @@ import "testing"
 // Create the users if required. The *md5* users have MD5 verifiers in
 // pg_authid, and *scram* users have SCRAM-SHA-256. pg_hba.conf is configured
 // (in .travis.sh) to authenticate *1 users with md5 and *2 with scram.
-func checkSCRAMSetup(t *testing.T) {
+func checkSCRAMSetup(t *testing.T) bool {
 	db := openTestConn(t)
 	defer db.Close()
+	if getServerVersion(t, db) < 100000 {
+		// skip the test if the postgres version is less than 10
+		return false
+	}
 	_, err := db.Exec(`
 		DROP USER IF EXISTS pqgomd5u1, pqgomd5u2, pqgoscramu1, pqgoscramu2;
 		SET password_encryption='md5';
@@ -20,10 +24,13 @@ func checkSCRAMSetup(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
+	return true
 }
 
 func testSCRAMSuccessCase(t *testing.T, conninfo string) {
-	checkSCRAMSetup(t)
+	if !checkSCRAMSetup(t) {
+		return
+	}
 	db, err := openTestConnConninfo(conninfo)
 	if err != nil {
 		t.Fatal(err)
@@ -42,14 +49,16 @@ func TestSCRAMBothScram(t *testing.T) {
 
 // user has SCRAM verifier, pg_hba picks md5 (should pass)
 func TestSCRAMUserScramServerMD5(t *testing.T) {
-	checkSCRAMSetup(t)
 	testSCRAMSuccessCase(t, `user=pqgoscramu1 password=se%r-*tpΣβ`)
 }
 
-// user has MD5 verifier, pg_hba picks scram (should fail)
+// user has MD5 verifier, pg_hba picks scram (should fail because
+// "user does not have a valid SCRAM identifier")
 func TestSCRAMUserMD5ServerScram(t *testing.T) {
-	checkSCRAMSetup(t)
-	db, err := openTestConnConninfo(`user=pqgomd5u2 password=incorrect`)
+	if !checkSCRAMSetup(t) {
+		return
+	}
+	db, err := openTestConnConninfo(`user=pqgomd5u2 password=se%r-*tpΣβ`)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -68,7 +77,9 @@ func TestSCRAMNoScram(t *testing.T) {
 
 // user has SCRAM verifier, pg_hba picks scram, but wrong password
 func TestSCRAMWrongPass(t *testing.T) {
-	checkSCRAMSetup(t)
+	if !checkSCRAMSetup(t) {
+		return
+	}
 	db, err := openTestConnConninfo(`user=pqgoscramu2 password=incorrect`)
 	if err != nil {
 		t.Fatal(err)
@@ -84,6 +95,11 @@ func TestSCRAMWrongPass(t *testing.T) {
 // allows it.
 func TestSCRAMStrangePasswords(t *testing.T) {
 	db := openTestConn(t)
+	if getServerVersion(t, db) < 100000 {
+		// skip the test if the postgres version is less than 10
+		db.Close()
+		return
+	}
 	_, err := db.Exec(`
 		DROP USER IF EXISTS pqgoscramu2;
 		SET password_encryption='scram-sha-256';
